make the default action to not signal parent process

+++ tools/pwrdaemon/pwrdaemon.cc
@@ -98,7 +98,10 @@ int main( int argc, char* argv[] )
 
 	rc = pthread_mutex_lock(&mutex);	
 	assert(0==rc);
-	kill(getppid(),SIGUSR1);
+
+    if ( getenv( "POWERRT_SIGNAL" ) ) {
+	    kill(getppid(),SIGUSR1);
+    }
 
 	while ( srvrThreads.size() ) {
 		//printf("wait for server thread to exit\n");
